Reworked folder structure, prepare place for auth subapp
